+++ MP/code/renderer/tr_noise.c
@@ -54,7 +54,7 @@ void R_NoiseInit( void ) {
 	}
 }
 
-float R_NoiseGet4f( float x, float y, float z, float t ) {
+float R_NoiseGet4f( float x, float y, float z, double t ) {
 	int i;
 	int ix, iy, iz, it;
 	float fx, fy, fz, ft;
